Calculation of the estimate for a price index

Calculation of a price index

Usage

CalcInd(data, baseVar, pVar, type, groupVar, consumVar = NULL, wVar = NULL)

Arguments

data

The dataset

baseVar

The variable name for the base goods price.

pVar

The variable name for the goods price.

type

The type of index to calculate. Choose from: 'carli', 'dutot', 'jevons'.

groupVar

The variable used for grouping observations - sometimes called the elementary group

consumVar

The variable used for grouping elementary groups. This may be the publishing level.

wVar

The variable used to weight the elementary groups up to the consumer publishing level.

Value

The estimates for the index are returned as a vector of length equal to the number of groups (in groupVar or consumVar if specified).

Examples

{
data(priceData)
# Calculate index for consumer groups
CalcInd(data = priceData, baseVar = "b1", pVar = "p1", groupVar = "varenr", wVar = "weight", 
  consumVar = "coicop", type = "dutot")
CalcInd(data = priceData, baseVar = "b1", pVar = "p1", groupVar = "varenr", wVar = "weight", 
  consumVar = "coicop", type = "carli")
CalcInd(data = priceData, baseVar = "b1", pVar = "p1", groupVar = "varenr", wVar = "weight", 
  consumVar = "coicop", type = "jevons")
  
# Calculate index for elemenatry groups (weighted)
CalcInd(data = priceData, baseVar = "b1", pVar = "p1", groupVar = "varenr", wVar = "weight", 
   type = "dutot")
   
   }
#> Warning: Elementary group weights did not add to one and have been scaled.
#> Warning: Elementary group weights did not add to one and have been scaled.
#> Warning: Elementary group weights did not add to one and have been scaled.
#> Warning: No consumer group variable was specified so an index was calculated for each elementary group.
#> Warning: Elementary group weights did not add to one and have been scaled.
#>           1           2           3           4           5           6 
#> 0.003690734 0.004574402 0.004312178 0.005857209 0.006229546 0.003796148 
#>           7           8           9          10          11          12 
#> 0.005153084 0.006806378 0.007038223 0.006247176 0.003427868 0.003782870 
#>          13          14          15          16          17          18 
#> 0.005440269 0.003859727 0.004616546 0.004029834 0.006042378 0.005364716 
#>          19          20          21          22          23          24 
#> 0.005560842 0.003951132 0.006994809 0.004817830 0.003648421 0.004598474 
#>          25          26          27          28          29          30 
#> 0.004736172 0.004557512 0.006590626 0.005661123 0.004434091 0.004606857 
#>          31          32          33          34          35          36 
#> 0.005558475 0.006602647 0.004339779 0.004432311 0.004291868 0.006359639 
#>          37          38          39          40          41          42 
#> 0.003863258 0.006173791 0.006242554 0.004355449 0.006728157 0.004259546 
#>          43          44          45          46          47          48 
#> 0.005208712 0.003546806 0.006274254 0.004540499 0.005805544 0.003443981 
#>          49          50          51          52          53          54 
#> 0.004056486 0.004175336 0.006476204 0.006258926 0.005534202 0.006210352 
#>          55          56          57          58          59          60 
#> 0.006618152 0.003656793 0.003360781 0.006008725 0.003794584 0.006602543 
#>          61          62          63          64          65          66 
#> 0.003934892 0.006085736 0.005737599 0.003894035 0.004561931 0.004387910 
#>          67          68          69          70          71          72 
#> 0.003897375 0.003537585 0.006608307 0.006052793 0.005248526 0.004511527 
#>          73          74          75          76          77          78 
#> 0.006581023 0.005461424 0.004460249 0.004756056 0.004177449 0.004927417 
#>          79          80          81          82          83          84 
#> 0.005566090 0.004767889 0.006056792 0.006546383 0.005053181 0.004906440 
#>          85          86          87          88          89          90 
#> 0.004924946 0.006291060 0.006244610 0.004641011 0.005255363 0.005486444 
#>          91          92          93          94          95          96 
#> 0.005247767 0.004477662 0.006652127 0.005547310 0.005722959 0.005063146 
#>          97          98          99         100         101         102 
#> 0.004185397 0.006823353 0.005741222 0.005640345 0.005321651 0.003897087 
#>         103         104         105         106         107         108 
#> 0.005215300 0.006341753 0.004436329 0.006049116 0.006595821 0.006632877 
#>         109         110         111         112         113         114 
#> 0.005001156 0.004147577 0.004924067 0.003526209 0.005366159 0.003960473 
#>         115         116         117         118         119         120 
#> 0.005571279 0.004767108 0.004642535 0.004501772 0.006457457 0.003746269 
#>         121         122         123         124         125         126 
#> 0.004005676 0.004481291 0.005508342 0.005946667 0.005031309 0.006178018 
#>         127         128         129         130         131         132 
#> 0.006683174 0.004079706 0.004092843 0.006364143 0.004945438 0.003830489 
#>         133         134         135         136         137         138 
#> 0.006050152 0.004398309 0.005240343 0.004189160 0.005619771 0.006053609 
#>         139         140         141         142         143         144 
#> 0.005666885 0.006314022 0.004400935 0.003645619 0.004169214 0.006033034 
#>         145         146         147         148         149         150 
#> 0.003457081 0.006921970 0.005886389 0.005281317 0.005846475 0.004763711 
#>         151         152         153         154         155         156 
#> 0.004254077 0.005731677 0.004002628 0.006169066 0.005777083 0.003647440 
#>         157         158         159         160         161         162 
#> 0.005309391 0.006285172 0.006426551 0.004035403 0.006084425 0.005298800 
#>         163         164         165         166         167         168 
#> 0.003435208 0.004608012 0.004047614 0.006055387 0.005365508 0.006811039 
#>         169         170         171         172         173         174 
#> 0.005561238 0.004085711 0.004333350 0.005630404 0.004617727 0.006394797 
#>         175         176         177         178         179         180 
#> 0.004186620 0.003871668 0.005736664 0.006474498 0.005425357 0.004964111 
#>         181         182         183         184         185         186 
#> 0.005265847 0.004745578 0.004647996 0.003865937 0.006755563 0.006130340 
#>         187         188         189         190         191         192 
#> 0.003802310 0.004182729 0.005012046 0.006215142 0.004842937 0.004138468 
#>         193         194         195         196         197         198 
#> 0.005324234 0.003662206 0.005652316 0.003510206 0.006025057 0.005479365 
#>         199         200 
#> 0.005797400 0.004649894